+++ src/91indexeddb.js
@@ -316,7 +316,7 @@ IDB.deleteFromTable = function(databaseid, tableid, wherefn,params, cb){
 //		  		console.log(222,event);
 //		  		console.log(333,cursor);
 		  	if(cursor) {
-		  		if(wherefn(cursor.value,params)) {
+		  		if((!wherefn) || wherefn(cursor.value,params)) {
 //		  		console.log(cursor);
 		  			cursor.delete();
 		  			num++;
@@ -331,4 +331,47 @@ IDB.deleteFromTable = function(databaseid, tableid, wherefn,params, cb){
 	}		
 }
 
+IDB.updateTable = function(databaseid, tableid, assignfn, wherefn, params, cb){
+	// console.log(arguments);
+	// console.trace();
+	var ixdbid = alasql.databases[databaseid].ixdbid;
+	var request = window.indexedDB.open(ixdbid);
+	request.onsuccess = function(event) {
+	  	var res = [];
+	  	var ixdb = event.target.result;
+//	  	console.log(444,ixdb, tableid, ixdbid);
+	  	var tx = ixdb.transaction([tableid], 'readwrite');
+	  	var store = tx.objectStore(tableid);
+	  	var cur = store.openCursor();
+	  	var num = 0;
+//	  	console.log(cur);
+	  	cur.onblocked = function(event) {
+//	  		console.log('blocked');
+	  	}
+	  	cur.onerror = function(event) {
+//	  		console.log('error');
+	  	}
+	  	cur.onsuccess = function(event) {
+//	  		console.log('success');
+		  	var cursor = event.target.result;
+//		  		console.log(222,event);
+//		  		console.log(333,cursor);
+		  	if(cursor) {
+		  		if((!wherefn) || wherefn(cursor.value,params)) {
+		  		//console.log(cursor);
+		  			var r = cursor.value;
+					assignfn(r,params);
+				//	console.log('update 363',r);
+		  			cursor.update(r);
+		  			num++;
+		  		}
+		  		cursor.continue();
+		  	} else {
+//		  		console.log(555, res,idx,query);
+		  		ixdb.close();
+		  		cb(num);
+		  	}
+	  	}
+	}		
+}