+++ test/test158.js
@@ -16,19 +16,24 @@ describe('Test 158 - INSERT/DELETE/UPDATE in IndexedDB', function() {
 			use test158; \
 			drop table if exists cities; \
 			create table cities (city string)",[], function(res) {
-				assert.deepEqual(res, [1,1,1,1,1]);
+			res[0] = 1;
+			assert.deepEqual(res, [1,1,1,1,1]);
 //				console.log(20);
-				alasql("insert into cities values ('Moscow'),('Paris'),('Minsk'),('Riga'),('Tallinn')",[],function(res){
+			alasql("insert into cities values ('Moscow'),('Paris'),('Minsk'),('Riga'),('Tallinn')",[],function(res){
 //				console.log(22);
-					assert(res,5);
-					alasql("select column * from cities where city like 'M%' order by city", [], function(res){
+				assert(res,5);
+				alasql("select column * from cities where city like 'M%' order by city", [], function(res){
 //				console.log(25, res);
-						assert.deepEqual(res,['Minsk','Moscow']);
-						alasql('delete from cities where city in ("Riga","Tallinn","Moscow")', [], function(res) {
-							assert(res == 3);
-							alasql('select column * from cities order by city', [], function(res) {
-								assert.deepEqual(res, ["Minsk","Paris"]);
-
+					assert.deepEqual(res,['Minsk','Moscow']);
+					alasql('delete from cities where city in ("Riga","Tallinn","Moscow")', [], function(res) {
+						assert(res == 3);
+						alasql('select column * from cities order by city', [], function(res) {
+							assert.deepEqual(res, ["Minsk","Paris"]);
+							alasql("update cities set city = 'Vilnius' where city = 'Minsk'", [], function(res){
+								assert(res == 1);
+								alasql('select column * from cities order by city', [], function(res) {
+									assert.deepEqual(res, ["Paris","Vilnius"]);
+//console.log(res);
 									alasql('detach database test158',[],function(res) {
 			//							console.log(52);
 										assert(res==1);
@@ -38,42 +43,15 @@ describe('Test 158 - INSERT/DELETE/UPDATE in IndexedDB', function() {
 											done();
 										});
 									});
-
 								});
 							});
 						});
 					});
-//				});
-		});
-
-	});
-if(false) {
-	it("2. UPDATE and DELETE", function(done){
-
-		alasql("update cities set city = 'Vilnius' where city = 'Minsk'", [], function(res){
-			assert(res == 1);
-			alasql('delete from cities where city in ("Riga","Tallinn","Moscow")', [], function(res) {
-				assert(res == 3);
-				alasql('select column * from cities order by city', [], function(res) {
-					assert.deepEqual(res, ["Berlin","Paris","Vilnius"]);
-					done();
 				});
 			});
 		});
 
 	});
-	it("99. Drop database", function(done){
-		alasql('detach database test158',[],function(res) {
-			console.log(52);
-			assert(res==1);
-			alasql('drop indexeddb database test158',[],function(res){
-				console.log(51,res);
-				assert(res==1);
-				done();
-			});
-		});
-	});
-}
 
 });