+++ src/84from.js
@@ -77,3 +77,88 @@ alasql.from.CSV = function(filename, opts, cb, idx, query) {
 	});
 	return res;
 };
+
+
+alasql.from.XLS = function(filename, opts, cb, idx, query) {
+	if(typeof exports === 'object') {
+		var XLSX = require('xlsjs');
+	} else {
+		var XLSX = window.XLS;
+		if(!XLSX) {
+			throw new Error('XLS library is not attached');
+		}
+	}
+	return XLSXLSX(XLSX,filename, opts, cb, idx, query);
+};
+
+alasql.from.XLSX = function(filename, opts, cb, idx, query) {
+	if(typeof exports === 'object') {
+		var XLSX = require('xlsx');
+	} else {
+		var XLSX = window.XLSX;
+		if(!XLSX) {
+			throw new Error('XLSX library is not attached');
+		}
+	}
+	return XLSXLSX(XLSX,filename, opts, cb, idx, query);
+};
+
+function XLSXLSX(XLSX,filename, opts, cb, idx, query) {
+	var opt = {};
+	if(!opts) opts = {};
+	alasql.utils.extend(opt, opts);
+	var res;
+
+	alasql.utils.loadBinaryFile(filename,!!cb,function(data){
+		var workbook = XLSX.read(data,{type:'binary'});
+//		console.log(workbook);
+		var sheetid;
+		if(typeof opt.sheetid == 'undefined') {
+			sheetid = workbook.SheetNames[0];
+		} else {
+			sheetid = opt.sheetid;
+		};
+		var range;
+		if(typeof opt.range == 'undefined') {
+			range = workbook.Sheets[sheetid]['!ref'];
+		} else {
+			range = opt.range;
+			if(workbook.Sheets[sheetid][range]) range = workbook.Sheets[sheetid][range];
+		};
+		var rg = range.split(':');
+		var col0 = rg[0].match(/[A-Z]+/)[0];
+		var row0 = rg[0].match(/[0-9]+/)[0];
+		var col1 = rg[1].match(/[A-Z]+/)[0];
+		var row1 = rg[1].match(/[0-9]+/)[0];
+//		console.log(114,rg,col0,col1,row0,row1);
+
+		var hh = {};
+		for(var j=alasql.utils.xlscn(col0);j<=alasql.utils.xlscn(col1);j++){
+			var col = alasql.utils.xlsnc(j);
+			if(opt.headers) {
+				hh[col] = workbook.Sheets[sheetid][col+""+row0].v;
+			} else {
+				hh[col] = col;
+			}
+		}
+		var res = [];
+		if(opt.headers) row0++;
+		for(var i=row0;i<=row1;i++) {
+			var row = {};
+			for(var j=alasql.utils.xlscn(col0);j<=alasql.utils.xlscn(col1);j++){
+				var col = alasql.utils.xlsnc(j);
+				if(workbook.Sheets[sheetid][col+""+i]) {
+					row[hh[col]] = workbook.Sheets[sheetid][col+""+i].v;
+				}
+			}
+			res.push(row);
+		}
+
+		if(cb) res = cb(res, idx, query);
+	}, function(err){
+		throw err;
+	});
+
+	return res;
+};
+